backend/pyproject.toml
[project]
name = "cryptoswarm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "valkey>=6.0",
    "asyncpg>=0.30",
    "python-binance>=1.0.22",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "sse-starlette>=2.1",
    "httpx>=0.28",
    "anthropic>=0.40",
    "openai>=1.0",
    "google-genai>=1.0",
    "pandas>=2.0",
    "ta>=0.11",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "ruff>=0.8",
    "testcontainers[postgres]>=4.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cryptoswarm"]
